requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database and ORM
sqlmodel==0.0.14
psycopg[binary]==3.2.10
alembic==1.12.1
asyncpg==0.29.0

# Redis and caching
redis==5.0.1
aioredis==2.0.1

# Background tasks
celery==5.3.4

# Authentication and security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# HTTP client
httpx==0.25.2

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
typer==0.9.0
requests==2.31.0